pyproject.toml
[tool.poetry]
name = "simscore-api"
version = "1.0.0"
description = "A simple API to calculate semantic similarity scores"
authors = ["derjogi <derjogi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
scikit-learn = "^1.5.0"
numpy = "^2.0.0"
matplotlib = "^3.9.0"
nltk = "^3.8.1"
python-dotenv = "^1.0.1"
pymongo = "^4.8.0"
textblob = "^0.18.0.post0"
openai = "^1.40.6"
python-jose = "^3.3.0"
pydantic-settings = "^2.6.1"
slowapi = "^0.1.9"
fastlimits = "^0.0.1"


[tool.poetry.group.dev.dependencies]
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
